refactor: improve code layering of trace-etl module
